Adding rail trips to "bus" scenario in get_scenario_settings
We are adding rail trips to bus scenario so it becomes a "public transport" scenario. The propensities are computed by summing them both and the reallocation of trips for this scenario are only for bus mode.
In create_max_mode_share_scenarios I commented what I did in the previous commit (sampling with replacement) because it's not needed anymore