Fixing issues in injuries
James pointed out that the difference between the number of deaths (absolute number) in the baseline and the cycling scenario was different by orders of magnitude between the health outcomes (run in summary_tables markdown) and the injuries results (run in injury_tables markdown).
I realized that deaths was being added twice for some modes, and this is what I fixed.