New Mexico City shape file
This better corresponds to the travel survey data and boundary. This shape file comes from OSM and it covers an area of 1499 km2. This is very close to 1,485 km2 which is the recorded offical area https://en.wikipedia.org/wiki/Mexico_City
Please use this file in any EDGAR calculation moving forward (15 June 2022)