Add files via upload
This better corresponds to the travel survey data and boundary. This shape file comes from OSM and it covers an area of 9,497 km2. This is very close to 9,459 km2 which is the recorded offical Metro area https://en.wikipedia.org/wiki/Belo_Horizonte
Please use this file in any EDGAR calculation moving forward (15 June 2022)